update chrome and node
there is a test failing - but it is passing locally.  locally I had to
update to chrome 117 - since the chromedriver for 116 (the current stable
version) is really hard to find - the stable channel is showing 117 even
though that's the same as beta - maybe they haven't quite gotten their
system down yet

also there are warnings about the old chrome version so I have changed
it to the one I am using locally.. I think it is old too.

I think this is an example of "bit rot"